requirements.txt
huggingface-hub
matplotlib-inline
numpy
tensorboard-data-server
tensorflow-estimator
tensorflow-io-gcs-filesystem
Django
djangorestframework
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
huggingface-hub
keras
matplotlib
pydantic
scikit-learn
seaborn
spacy
tensorflow
transformer
uvicorn
fastapi